PromotionEngine/Controllers/ValuesController.cs
using BusinessLayer;
using PromotionEngine.Entity;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Net;
using System.Net.Http;
using System.Web.Http;

namespace PromotionEngine.Controllers
{
    public class ValuesController : ApiController
    {
        // GET api/values
        public IEnumerable<string> Get()
        {
            return new string[] { "value1", "value2" };
        }

        // GET api/values/5
        public string Get(int id)
        {
            return "value";
        }

        // POST api/values
        public string Post([FromBody]Sku sku)
        {
            var cost = "0";
            try
            {
                Calculation calulation = new Calculation();

                cost = calulation.GetCost(sku);
                return cost;
            }
            catch (Exception ex)
            {
                throw ex;
            }
            
        }

        // PUT api/values/5
        public void Put(int id, [FromBody]string value)
        {
        }

        // DELETE api/values/5
        public void Delete(int id)
        {
        }
    }
}